requirements.txt
fastapi>=0.68.0,<0.69.0
aiofiles
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
requests
pipe
python-dotenv
openai

